occurrences/requirements.txt
Django==1.11.6
djangorestframework==3.9.2
psycopg2==2.6.2
